+++ environment.yml
@@ -10,6 +10,7 @@ dependencies:
   - graphene
   - iso8601
   - jupyterlab >=0.35,<0.36
+  - nodejs >=8,<9
   - pandas
   - pyld
   - python >=3.6,<3.7
@@ -17,7 +18,6 @@ dependencies:
   - requests
   - requests-cache
   - werkzeug
-  - nodejs >=8,<9
   - pip:
     - graphql-ws
     - graphene-tornado
