+++ package.json
@@ -13,6 +13,7 @@
   "scripts": {
     "bootstrap": "jlpm --no-optional && jlpm lerna bootstrap && jlpm build",
     "build": "lerna run build --stream",
+    "watch": "lerna run --stream --parallel --prefix watch",
     "prettier": "prettier --write *.json *.md *.yml && lerna exec --stream -- prettier --write --config=../../.prettierrc ./src/**/*.ts ./style/**/*.css ./*.json ./*.md"
   }
 }
