+++ setup.cfg
@@ -31,7 +31,7 @@ install_requires =
     notebook
     werkzeug
 package_dir =
-    = src/py
+    = src
 packages = find:
 include_package_data = True
 zip_safe = False
@@ -44,7 +44,7 @@ jupyter_graphql_schema =
 
 [options.packages.find]
 where =
-    src/py
+    src
 
 [flake8]
 max-line-length = 88
