+++ src/jupyter_graphql/constants.py
