+++ src/jupyter_graphql/manager.py
