+++ src/jupyter_graphql/schema/contents/base.py
