+++ src/jupyter_graphql/subscriptions.py
