+++ src/jupyter_graphql/templates/graphiql.html
